requirements.txt
Flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.1.0